SysUI: Use VectorDrawable's for rendering battery
The current implementation of BatteryMeterView uses arrays of points
that are drawn programmatically.  This limits the ability to overlay
or theme the battery.  By utilizing VectorDrawables and
AnimatedVectorDrawables we can achieve the same look and feel of
stock while allowing for more unique battery meters.

Forward port the AVD battery commits from CM 13.0

Change-Id: I91b4f5229774c979a7aa9ab771ffcb95ceb71b84