Battery icon improvements
1) on stock aosp, for landscape and portrait battery icon mode,
pct and bolt are white when the battery level is less or equal
to 30%, and this is good because then they are more visible in
the battery icon.

2) move back the critical level symbol ! out of the pctText related
if else so it will be always displayed in the icon when the level is
<= mCriticalLevel (it's good to warn the user that level is critical
also if he doesn't want pct inside the icon)

3) change the boltOverlay bool name to an easy-to-understand one

4) on landscape or portrait icon mode, when level is <=15, the pct
inside the icon is red. Before this commit, though, the XOR porterduff
mode was applied, rather now let's do what we did in the point (1) of
this commit for the white pct: make the pct full red and don't care of
statusbar colorization, so the pct will be more visible.
It will be full white if power saver enabled because getColorForLever
provides the needed code.

5) If the user sets no pct inside the icon, show a "+" symbol inside
the icon when power save mode is enabled.
PS: made a vector for the "+", the string looks like shit

6) remove aosp leftovers we never used

7) code cleanup

8)unlink the pct/charge level red color from the powersave mode color
so we can use a different one. Set it to #E53935 (material red 600) so
it's better visible on the dark grey frame color

Change-Id: I65fc9d2fa46310ec17d12b26aba6d9595fea3dc6